rename(*): Add pages folder
