+++ src/pages/Todo.css
@@ -1,0 +1,120 @@
+@charset "utf-8";
+
+/* Input */
+.add-form-box {
+  display: flex;
+  padding: 26px 32px;
+  border-radius: 16px;
+  background: #f7f7f7;
+}
+
+.add-form-box label {
+  flex: 1 1;
+  display: flex;
+  align-items: center;
+}
+
+.form {
+  box-sizing: border-box;
+  font-size: 16px;
+  border: 1px solid #cdcdcd;
+  margin: 0 28px 0 12px;
+  padding: 0 10px;
+  height: 48px;
+  border-radius: 6px;
+  width: auto;
+  min-width: 150px;
+  background-color: #fff;
+}
+
+.add-form-box .form {
+  flex: 1;
+}
+
+.add-form-box button {
+  width: 120px;
+  height: 48px;
+  margin-left: auto;
+  border: 0;
+  border-radius: 6px;
+  padding: 0 16px;
+  /* background: #203a91; */
+  background: #005955;
+  color: #fff;
+}
+
+/* TodoList */
+h2 {
+  margin: 56px 0 16px;
+  font-size: 22px;
+  font-weight: 600;
+}
+
+.todos-list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px 2%;
+}
+.todos-list > li {
+  overflow: hidden;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  width: 32%;
+  min-height: 220px;
+  margin-bottom: 10px;
+  padding: 30px 24px 94px;
+  border: 3px solid #203a91;
+  border: 3px solid #005955;
+  border-radius: 16px;
+  box-shadow: 0 8px 15px rgba(11, 51, 82, 0.2);
+}
+
+.todos-list > li > strong {
+  display: block;
+  margin: 0 0 14px;
+  font-size: 24px;
+  font-weight: 600;
+}
+.todos-list > li > p {
+  flex: 1;
+  color: #888;
+}
+
+.todos-list .btn-wrap {
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  margin: 0;
+  border-radius: 0 0 10px 10px;
+  border-radius: 0 0 10px 10px;
+  overflow: hidden;
+}
+.todos-list .btn-wrap button {
+  float: left;
+  min-width: 50%;
+  height: 54px;
+  margin: 0;
+  border-radius: 0;
+  box-sizing: border-box;
+  border: 0;
+  background: #f2f2f2;
+}
+
+.todos-list .btn-wrap .btn-done {
+  color: #fff;
+  background: #203a91;
+  background: #005955;
+}
+
+.todos-list.done > li {
+  box-shadow: none;
+  border-color: #eaebed;
+  border-color: #dbdde1;
+}
+
+.todos-list.done .btn-wrap .btn-done {
+  background: #fbb938;
+}
